requirements.txt
# FastAPI backend
fastapi
uvicorn
pydantic

# Google GenAI SDK (Gemini client)
google-genai

# Frontend / client
streamlit
requests

# Env handling
python-dotenv
